minor: split up TestOperation.swift into separate files
